Overhaul of stage-in and process inputs.
The stage-in parsing script may now parse lists of URLs that use the same credentials.

The stage-in parsing script now parses a JSON of the inputs instead of cluttering the command line.

The process step now always treats stage-in scripts as an array of dicts describing files.

The process step now loads a JSON from /tmp/inputs.json to set its papermill parameters.